Alterando estrutura do app para padrão Next e separando págs login-cadastro
