+++ components/auth/AuthDetails.jsx
@@ -1,6 +1,6 @@
 import { onAuthStateChanged, signOut } from 'firebase/auth';
 import React, { useEffect, useState } from 'react';
-import { auth } from '../firebaseConfig';
+import { auth } from '../../firebase/firebaseConfig';
 
 const AuthDetails = () => {
     const [authUser, setAuthUser] = useState(null);
