+++ package.json
@@ -1,6 +1,10 @@
 {
   "dependencies": {
     "firebase": "^10.7.0",
+    "next": "^14.0.3",
     "react-router-dom": "^6.20.1"
+  },
+  "scripts": {
+    "dev": "next dev"
   }
 }
