+++ pages/auth/login.jsx
@@ -1,0 +1,14 @@
+import SignIn from "../../components/auth/SignIn"
+import AuthDetails from "../../components/auth/AuthDetails";
+
+export default function Login() {
+    return (
+        <div className="App">
+            <div className="App-logo">
+                <span>Bem-vindo ao SimpleTask</span>
+            </div>
+            <SignIn />
+            <AuthDetails />
+        </div>
+    );
+}
