+++ pages/index.jsx
@@ -1,0 +1,8 @@
+export default function Home(){
+    return(
+        <div>
+            <a href="/auth/login">Login</a>
+            <a href="/auth/cadastro">Cadastro</a>
+        </div>
+    );
+}
