--- react-firebase-auth/src/App.css
@@ -1,33 +1,0 @@
-/* App.css */
-
-body {
-  font-family: Arial, sans-serif;
-  margin: 0;
-  padding: 0;
-}
-
-.App {
-  text-align: center;
-  padding: 20px;
-  background-color: #3498db;
-  color: #ffffff;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  height: 100vh;
-}
-
-.App-logo {
-  margin-bottom: 20px;
-}
-
-.App-logo img {
-  width: 100px; /* Ajuste a largura conforme necessário */
-  height: auto;
-}
-
-.App-logo span {
-  font-size: 1.5em;
-  margin-top: 10px;
-}
