--- react-firebase-auth/src/pages/Home.jsx
@@ -1,1 +1,0 @@
-
